+++ src/main/java/com/sidgames5/betterdeco/block/ModBlocks.java
@@ -18,6 +18,7 @@ import net.minecraft.util.Identifier;
 public class ModBlocks {
     // Blocks
     public static final Block PAPER_REAM = registerBlock("paper_ream", new Block(FabricBlockSettings.of(Material.CARPET).strength(0.0F, 0.0F)), ItemGroups.COLORED_BLOCKS);
+    public static final Block PLASTIC_BLOCK = registerBlock("plastic_block", new Block(FabricBlockSettings.of(Material.CARPET).strength(1.0F, 1.0F)), ItemGroups.BUILDING_BLOCKS);